refact: pages moved to src dir
