+++ src/components/index.ts
@@ -2,3 +2,6 @@ export * from './layouts/MainLayout'
 export * from './PostCard'
 export * from './PostContainer'
 export * from './Providers'
+export * from './FormInputs'
+export * from './RoomCard'
+export * from './RoomGrid'
