--- utils/pageProps.ts
@@ -1,6 +1,0 @@
-import { AppProps } from 'next/app'
-import { PageLayout } from '@utils/pageLayout'
-
-export type pageProps = AppProps & {
-  Component: PageLayout
-}
